factoring 2
git-svn-id: https://kunghit.ugrad.cs.ubc.ca:8411/repos/2013w2/users/j6j7@1465 07e144b1-f7ef-4d70-afd9-7d896f8a2c5e

+++ minijava/src/codegen/x86_64/X86_64Muncher.java
@@ -79,7 +79,7 @@ public class X86_64Muncher extends Muncher {
     final Pat<Integer> _scale_ = new Wildcard<Integer>() {
       @Override
       public void match(Integer toMatch, Matched matched) throws Failed {
-        int value = (Integer) toMatch;
+        int value = toMatch;
         if (value == 1 || value == 2 || value == 4 || value == 8)
           super.match(toMatch, matched);
         else
